+++ ansible/test_plugins/functional.py
@@ -1,0 +1,27 @@
+# Copyright (c) 2017 StackHPC Ltd.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"); you may
+# not use this file except in compliance with the License. You may obtain
+# a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations
+# under the License.
+
+
+def is_in(item, container):
+    """Tests whether an item is in a container (e.g. a list)."""
+    return item in container
+
+
+class TestModule(object):
+    """'Functional' programming tests."""
+
+    def tests(self):
+        return {
+            'is_in': is_in,
+        }