+++ doc/source/release-notes.rst
@@ -17,6 +17,8 @@ Features
 * Adds support for configuration of custom fluentd filters, and additional
   config file templates for heat, ironic, keystone, magnum, murano, sahara, and
   swift in ``$KAYOBE_CONFIG_PATH/kolla/config/<component>/``.
+* Adds the command ``kayobe network connectivity check`` which can be used to
+  verify network connectivity in the cloud hosts.
 
 Upgrade Notes
 -------------